refactor(compile): hoist block-scope-rename machinery into StateMachineExitRoutingEmitter
The #711/#766 rename-then-delegate overrides (EmitConstDeclaration,
EmitCompoundAssign, EmitLogicalAssign, EmitPrefixIncrement,
EmitPostfixIncrement) plus RenameToken and the NoRenames sentinel were
byte-identical across IteratorMoveNextEmitter, AsyncMoveNextEmitter,
and AsyncArrowMoveNextEmitter. They now live once on the shared
StateMachineExitRoutingEmitter ancestor; each family supplies only its
BlockScopeRenames source. No intermediate class overrode these methods,
so virtual dispatch (and the emitted IL) is unchanged.

EmitterSyncTests allowlist moved accordingly.
