+++ Compilation/AsyncArrowMoveNextEmitter.Variables.cs
@@ -8,11 +8,8 @@ public partial class AsyncArrowMoveNextEmitter
 {
     // Per-binding storage names for block-scoped let/const shadows (#766), shared with the analyzer via
     // the analysis. Empty for the common no-shadow case (and for expression-bodied arrows).
-    private static readonly IReadOnlyDictionary<object, string> NoRenames = new Dictionary<object, string>();
-    private IReadOnlyDictionary<object, string> BlockScopeRenames => _analysis.BlockScopeRenames ?? NoRenames;
-
-    private static Token RenameToken(Token original, string lexeme) =>
-        new(original.Type, lexeme, original.Literal, original.Line, original.Start);
+    // The rename-then-delegate operator overrides that consume this live on StateMachineExitRoutingEmitter.
+    protected override IReadOnlyDictionary<object, string> BlockScopeRenames => _analysis.BlockScopeRenames ?? NoRenames;
 
     protected override void EmitVariable(Expr.Variable v)
     {
@@ -499,44 +496,6 @@ public partial class AsyncArrowMoveNextEmitter
         _il.Emit(OpCodes.Ldfld, _builder.FunctionDCField!);
     }
 
-    // Const declarations, compound/logical assignment, and increment/decrement reach the variable
-    // through the operator node's name token (or the increment operand). Rewriting that token to the
-    // shadowing binding's storage name before delegating to the base routes both the read and the
-    // write to the right field/local (#766). The base re-enters EmitVarDeclaration / EmitVariable /
-    // EmitStoreVariable (all of which resolve by name here), so the rename flows through consistently.
-
-    protected override void EmitConstDeclaration(Stmt.Const c)
-    {
-        if (BlockScopeRenames.TryGetValue(c, out var renamed))
-            c = c with { Name = RenameToken(c.Name, renamed) };
-        base.EmitConstDeclaration(c);
-    }
-
-    protected override void EmitCompoundAssign(Expr.CompoundAssign ca)
-    {
-        if (BlockScopeRenames.TryGetValue(ca, out var renamed))
-            ca = ca with { Name = RenameToken(ca.Name, renamed) };
-        base.EmitCompoundAssign(ca);
-    }
-
-    protected override void EmitLogicalAssign(Expr.LogicalAssign la)
-    {
-        if (BlockScopeRenames.TryGetValue(la, out var renamed))
-            la = la with { Name = RenameToken(la.Name, renamed) };
-        base.EmitLogicalAssign(la);
-    }
-
-    protected override void EmitPrefixIncrement(Expr.PrefixIncrement pi)
-    {
-        if (pi.Operand is Expr.Variable v && BlockScopeRenames.TryGetValue(v, out var renamed))
-            pi = pi with { Operand = v with { Name = RenameToken(v.Name, renamed) } };
-        base.EmitPrefixIncrement(pi);
-    }
-
-    protected override void EmitPostfixIncrement(Expr.PostfixIncrement poi)
-    {
-        if (poi.Operand is Expr.Variable v && BlockScopeRenames.TryGetValue(v, out var renamed))
-            poi = poi with { Operand = v with { Name = RenameToken(v.Name, renamed) } };
-        base.EmitPostfixIncrement(poi);
-    }
+    // The rename-then-delegate overrides for const declarations, compound/logical assignment, and
+    // increment/decrement live on StateMachineExitRoutingEmitter (shared with the iterator family).
 }
